open page timing done
have not tested this function yet but it should not need much adjustment. its pretty much good to go

+++ src/parser.c
@@ -74,7 +74,7 @@ void parser_next_line(Parser_t *parser) {
 
 MemoryRequest_t parse_line(char *line) {
   uint64_t time, address;
-  unsigned core, operation;
+  uint8_t core, operation;
   MemoryRequest_t memory_request;
 
   sscanf(line, "%llu %u %u %llx", &time, &core, &operation, &address);
